Combined both telemetry event collections to a single thread
